stop localizing @INC during module load for TIE
This was added as part of a bulk patch in perl5.git back
during the dot-in-@INC hysteria, which (as the commit message says)
was intended to protect optional module loads.

Except the module load during TIE is unconditional. Not optional.
The require() may have been wrapped in an eval{} but any exception
it throws is always rethrown. The reason for the eval{} is not to
be able to ignore module load errors.

In fact there are appears to be no good reason for the eval{} at all.
It serves no purpose other than to allow wrapping the error message…
and not just prefixed, but literally wrapped, with something appended
to the end, but without clear bracketing of the wrapped message.

The simplest alternative which addresses all these issues is to just
remove the eval{} entirely.