clean up a few more error messages
The croak removed in the last commit which added text after $@ was not
the only place where the module was propagating someone else’s error
with stuff appended to the end: a few lines away, $! was getting the
same treatment.

While we’re at, the last occurrence of this “; aborting” suffix can go
as well, just for consistency, even though that error message does not
follow the same pattern.