make the tests strict-clean non-invasively
This patch does not edit any existing lines, it only adds lines,
in order to preserve blame for now. This makes for unnatural code,
especially when it comes to making barewords pass, but many of the
tests would benefit from refactoring anyway, at which point blame
will be reassigned appropriately.