fix $Verbose typo in Memoize::NBDM_File
The typo also used to exist in Memoize::SDBM_File until last commit.

This patch may perhaps be read as a commentary of sorts on
https://perl.plover.com/yak/12views/samples/notes.html#sl-31
Does mindlessly turning on strictures replace thinking? Of course
it does not, but what it does do is catch stupid mistakes; mistakes
of a kind not worth the expenditure in attention and thinking that
their prevention and correction require, even of Mark Jason Dominus
himself.

    Civilization advances by extending the number of important
    operations which we can perform without thinking of them.
					Alfred North Whitehead