separate data from metadata in Memoize::Expire
This allows storing both in the same hash without stringifying the
values. This has a two-fold effect:

* If storage is backed by a plain hash, values can be references.
* Storage can still be backed by a plain DBM without serialisation so
  long as the values are simple scalars.

Storing data and metadata together under a single key cannot fulfil
both of these requirements at the same time, because then either the
data and metadata have to be concatenated into a string that is parsed
back apart on every access (as previously implemented) – making it
impossible to store references as values –, or else they must always be
stored as a data structure – entirely ruling out the use of a plain DBM
for persistence.

Keeping the metadata separate from the data avoids this dilemma.

This addresses https://rt.cpan.org/Ticket/Display.html?id=38344