Revert "simplify Memoize::AnyDBM_File"
The simplification of Memoize::AnyDBM_File in 1.06 was a mistake.

It turns out that while the old interface of Memoize::AnyDBM_File is
silly, the reason it is like that is that it (of course) emulates the
silly interface of AnyDBM_File. And that interface also constitutes the
exposed and documented API of that module. So emulating the silliness is
a must.

This is particularly unfortunate given the implementation of AnyDBM_File
as well as the purpose of Memoize::AnyDBM_File, which together mean that
it is not possible to implement Memoize::AnyDBM_File as a wrapper around
AnyDBM_File.

This reverts commit 4a41a7f9f3d2f0b0ac3a39c2df4211e80d92a298.