get rid of wrapper around _memoizer
Instead of using `eval` to generate a wrapper that calls _memoizer,
where the bulk of the logic resides, there is now a _wrap function
which returns a closure containing the logic directly.

This gets rid of one extra function call every time a memoized function
is called.

It also gets rid of the extra stack frame for the wrapper on threaded
perls (which these days is most perls, regrettably). This removes the
need to load Config.pm to tell threaded and unthreaded perls from each
other – though at the cost of loading Scalar::Util to set the prototype
on the closure.

So this trades one dependency for another. However, Scalar::Util takes
up less memory than Config in active use, and is also much more likely
than Config to already be loaded by some other part of the program, in
which case loading it is virtually free.

Using a closure also present an opportunity to convert all the $info
hash lookups in the wrapper logic to captured lexicals, for cleaner and
faster code. Upcoming commits will realise this opportunity.