drop t/expfile.t and t/lib/ExpireFile.pm
The test only really tests t/lib/ExpireFile.pm. This previously made
sense because it was included as Memoize::ExpireFile. Except it made
no sense even then because it already did nothing much useful and so
was also documented as only meant for use by the test suite. That is
to say this test is, as it has only ever been, pointless.