rewrite t/correctness.t and drop t/speed.t
The rewritten t/correctness.t directly checks the impact of memoization
on the amount of calls to the original naïve recursive function. This
establishes the same fact that the performance test in t/speed.t proves
but without the need to run the test for a significant length of time.