stop including article.html
Files included in a module distribution tarball that are outside the
structure expected by the CPAN toolchain are effectively invisible
nowadays. Basically nobody downloads, much less inspects, tarballs from
CPAN manually.

People do look at such contents in CPAN and Git web interfaces, but
those go out of their way to disallow viewing HTML files as a web pages
rather than as HTML source (and with good reason; however unfortunate).

The sum total is that it is *far* easier to read this article on Mark’s
site than to read (or even notice) the copy included in the tarball.