protect @_ from normalizer
The change in f2e9814c913865cccd51270fbee0324072351cb6 speeds up the
normalizer call by exposing the wrapper’s @_ to the normalizer, which
saves a bit of work during the function call but also means that any
existing normalizers which mutate their @_ now affect the @_ seen by
the memoized function. This is a regression.

As a capability the ability of the normalizer to affect the memoized
function’s @_ does not really make sense since it means the behaviour
of a function depends on whether it is memoized. In that regard it makes
more sense to protect @_ against changes by the normalizer.

Too bad for the performance gain.